+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/Config.java
@@ -149,8 +149,6 @@ public interface Config {
     }
 
     interface WirelessGridEntry {
-        boolean getUseEnergy();
-
         long getEnergyCapacity();
 
         long getOpenEnergyUsage();