+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/energy/CreativeEnergyStorage.java
@@ -1,9 +1,14 @@
-package com.refinedmods.refinedstorage2.api.network.impl.energy;
+package com.refinedmods.refinedstorage2.platform.common.internal.energy;
 
 import com.refinedmods.refinedstorage2.api.core.Action;
 import com.refinedmods.refinedstorage2.api.network.energy.EnergyStorage;
 
-public class InfiniteEnergyStorage implements EnergyStorage {
+public class CreativeEnergyStorage implements EnergyStorage {
+    public static final EnergyStorage INSTANCE = new CreativeEnergyStorage();
+
+    private CreativeEnergyStorage() {
+    }
+
     @Override
     public long getStored() {
         return Long.MAX_VALUE;