+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/PlatformImpl.java
@@ -3,18 +3,16 @@ package com.refinedmods.refinedstorage2.platform.forge;
 import com.refinedmods.refinedstorage2.api.core.Action;
 import com.refinedmods.refinedstorage2.api.grid.view.GridResourceFactory;
 import com.refinedmods.refinedstorage2.api.network.energy.EnergyStorage;
-import com.refinedmods.refinedstorage2.api.network.impl.energy.InfiniteEnergyStorage;
 import com.refinedmods.refinedstorage2.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
 import com.refinedmods.refinedstorage2.platform.common.AbstractPlatform;
 import com.refinedmods.refinedstorage2.platform.common.Config;
-import com.refinedmods.refinedstorage2.platform.common.block.ControllerType;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.transfer.TransferManager;
 import com.refinedmods.refinedstorage2.platform.common.util.BucketAmountFormatting;
 import com.refinedmods.refinedstorage2.platform.common.util.CustomBlockPlaceContext;
 import com.refinedmods.refinedstorage2.platform.forge.containermenu.ContainerTransferDestination;
-import com.refinedmods.refinedstorage2.platform.forge.integration.energy.ControllerForgeEnergy;
+import com.refinedmods.refinedstorage2.platform.forge.internal.energy.EnergyStorageAdapter;
 import com.refinedmods.refinedstorage2.platform.forge.internal.grid.ItemGridInsertionStrategy;
 import com.refinedmods.refinedstorage2.platform.forge.internal.grid.view.ForgeFluidGridResourceFactory;
 import com.refinedmods.refinedstorage2.platform.forge.internal.grid.view.ForgeItemGridResourceFactory;
@@ -174,21 +172,6 @@ public final class PlatformImpl extends AbstractPlatform {
     }
 
     @Override
-    public EnergyStorage createEnergyStorage(final ControllerType controllerType, final Runnable listener) {
-        return switch (controllerType) {
-            case NORMAL -> new ControllerForgeEnergy(listener);
-            case CREATIVE -> new InfiniteEnergyStorage();
-        };
-    }
-
-    @Override
-    public void setEnergy(final EnergyStorage energyStorage, final long stored) {
-        if (energyStorage instanceof ControllerForgeEnergy controllerForgeEnergy) {
-            controllerForgeEnergy.setSilently(stored);
-        }
-    }
-
-    @Override
     public TransferManager createTransferManager(final AbstractContainerMenu containerMenu) {
         return new TransferManager(containerMenu, ContainerTransferDestination::new);
     }
@@ -336,5 +319,13 @@ public final class PlatformImpl extends AbstractPlatform {
             y,
             DefaultTooltipPositioner.INSTANCE
         );
+    }
+
+    @Override
+    public Optional<EnergyStorage> getEnergyStorage(final ItemStack stack) {
+        return stack.getCapability(ForgeCapabilities.ENERGY)
+            .filter(EnergyStorageAdapter.class::isInstance)
+            .map(EnergyStorageAdapter.class::cast)
+            .map(EnergyStorageAdapter::getEnergyStorage);
     }
 }